+++ server/src/loaders/index.ts
@@ -1,10 +1,10 @@
 import { Application } from 'express';
-import envConfig from '@/config/env-config';
+import dotEnvConfig from '@/config/dot-env-config';
 import expressLoader from './express-loader';
 import dbLoader from './db-loader';
 
 const loaders = async (app: Application): Promise<void> => {
-  envConfig();
+  dotEnvConfig();
   console.log(`NODE_ENV=${process.env.NODE_ENV}, dotenv success`);
   expressLoader(app);
   console.log('express load');
